# backend/openapi.yaml
openapi: 3.0.3
info:
  title: BoxMeOut API
  version: 0.2.0
  description: REST API for the BoxMeOut decentralized boxing prediction market.
servers:
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    oracleKey:
      type: apiKey
      in: header
      name: X-Oracle-Key
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
...
    Market:
      type: object
      properties:
        id:               { type: integer }
        market_id:        { type: string }
        contract_address: { type: string }
        match_id:         { type: string }
        fighter_a:        { type: string }
        fighter_b:        { type: string }
        weight_class:     { type: string }
        title_fight:      { type: boolean }
        venue:            { type: string }
        scheduled_at:     { type: string, format: date-time }
        status:
          type: string
          enum: [open, locked, resolved, cancelled, disputed]
        outcome:
          type: string
          nullable: true
          enum: [fighter_a, fighter_b, draw, no_contest]
        pool_a:           { type: string }
        pool_b:           { type: string }
        pool_draw:        { type: string }
        total_pool:       { type: string }
        fee_bps:          { type: integer }
        resolved_at:      { type: string, format: date-time, nullable: true }
        oracle_used:      { type: string, nullable: true, enum: [primary, fallback, admin] }
        created_at:       { type: string, format: date-time }
        updated_at:       { type: string, format: date-time }
        ledger_sequence:  { type: integer }
    MarketOdds:
      type: object
      properties:
        market_id:   { type: string }
        odds_a:      { type: number }
        odds_b:      { type: number }
        odds_draw:   { type: number }
        pool_a:      { type: string }
        pool_b:      { type: string }
        pool_draw:   { type: string }
        total_pool:  { type: string }
    MarketStats:
      type: object
      properties:
        market_id:        { type: string }
        total_bets:       { type: integer }
        unique_bettors:   { type: integer }
        largest_bet_xlm:  { type: number }
        average_bet_xlm:  { type: number }
        total_pooled_xlm: { type: number }
    SimulatePayout:
      type: object
      properties:
        market_id:       { type: string }
        amount:          { type: number }
        outcome:         { type: integer, enum: [0, 1] }
        estimated_payout: { type: number }
        fee_xlm:         { type: number }
        odds:            { type: number }
...
    Bet:
      type: object
      properties:
        id:              { type: integer }
        market_id:       { type: string }
        bettor_address:  { type: string }
        side:
          type: string
          enum: [fighter_a, fighter_b, draw]
        amount:          { type: string }
        amount_xlm:      { type: number }
        placed_at:       { type: string, format: date-time }
        claimed:         { type: boolean }
        claimed_at:      { type: string, format: date-time, nullable: true }
        payout:          { type: string, nullable: true }
        tx_hash:         { type: string }
        ledger_sequence: { type: integer }
...
    Portfolio:
      type: object
      properties:
        address:          { type: string }
        active_bets:
          type: array
          items: { $ref: '#/components/schemas/Bet' }
        past_bets:
          type: array
          items: { $ref: '#/components/schemas/Bet' }
        pending_claims:
          type: array
          items: { $ref: '#/components/schemas/Bet' }
        total_staked_xlm: { type: number }
        total_won_xlm:    { type: number }
        total_lost_xlm:   { type: number }
...
paths:
...
  /api/markets:
    get:
      summary: List markets
      operationId: listMarkets
      tags: [Markets]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, locked, resolved, cancelled, disputed]
        - name: weight_class
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated list of markets
          content:
            application/json:
              schema:
                type: object
                properties:
                  markets:
                    type: array
                    items: { $ref: '#/components/schemas/Market' }
                  total: { type: integer }
                  page:  { type: integer }
                  limit: { type: integer }
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/markets/{market_id}:
    get:
      summary: Get market by ID
      operationId: getMarket
      tags: [Markets]
      parameters:
        - name: market_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Market detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Market' }
        '404':
          description: Market not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/markets/{market_id}/bets:
    get:
      summary: Get bets for a market
      operationId: getMarketBets
      tags: [Markets]
      parameters:
        - name: market_id
          in: path
          required: true
          schema: { type: string }
        - name: address
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated list of bets
          content:
            application/json:
              schema:
                type: object
                properties:
                  bets:
                    type: array
                    items: { $ref: '#/components/schemas/Bet' }
                  total: { type: integer }
                  page:  { type: integer }
                  limit: { type: integer }
        '404':
          description: Market not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/markets/{market_id}/odds:
    get:
      summary: Get current odds for a market
      operationId: getMarketOdds
      tags: [Markets]
      parameters:
        - name: market_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Current odds
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketOdds' }
        '404':
          description: Market not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/markets/{market_id}/stats:
    get:
      summary: Get aggregate stats for a market
      operationId: getMarketStats
      tags: [Markets]
      parameters:
        - name: market_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Market statistics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketStats' }
        '404':
          description: Market not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/markets/{market_id}/simulate:
    get:
      summary: Simulate payout for a given bet amount and outcome
      operationId: simulatePayout
      tags: [Markets]
      parameters:
        - name: market_id
          in: path
          required: true
          schema: { type: string }
        - name: amount
          in: query
          required: true
          schema: { type: number }
        - name: outcome
          in: query
          required: true
          schema: { type: integer, enum: [0, 1] }
      responses:
        '200':
          description: Simulated payout
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SimulatePayout' }
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/markets/{market_id}/resolve:
    post:
      summary: Resolve a market (admin only)
      operationId: resolveMarket
      tags: [Markets]
      security:
        - bearerAuth: []
      parameters:
        - name: market_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [winningOutcome]
              properties:
                winningOutcome: { type: integer, enum: [0, 1] }
      responses:
        '200':
          description: Market resolved
          content:
            application/json:
              schema:
                type: object
                properties:
                  tx_hash: { type: string }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Market not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ---------------------------------------------------------------------------
  # Bets
  # ---------------------------------------------------------------------------
  /api/bets/{bettor_address}:
    get:
      summary: Get all bets by a Stellar address
      operationId: getBetsByAddress
      tags: [Bets]
      parameters:
        - name: bettor_address
          in: path
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated list of bets
          content:
            application/json:
              schema:
                type: object
                properties:
                  bets:
                    type: array
                    items: { $ref: '#/components/schemas/Bet' }
                  total: { type: integer }
                  page:  { type: integer }
                  limit: { type: integer }
        '400':
          description: Invalid Stellar address
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /api/portfolio/{address}:
    get:
      summary: Get portfolio summary for a Stellar address
      operationId: getPortfolio
      tags: [Bets]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Portfolio summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Portfolio' }
        '400':
          description: Invalid Stellar address
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...